+++ CascableCore.xcframework/xros-arm64/CascableCore.framework/Headers/CBLCoreVersions.h
@@ -1,4 +1,4 @@
 /// Returns a detailed version string for the SDK.
-static NSString * const CBLCoreVerboseVersion NS_SWIFT_NAME(CascableCoreVerboseVersion) = @"15.0.2-gca678376";
+static NSString * const CBLCoreVerboseVersion NS_SWIFT_NAME(CascableCoreVerboseVersion) = @"16.0.0-g2f418540";
 /// Returns the build number for the SDK.
-static NSString * const CBLCoreBundleVersion NS_SWIFT_NAME(CascableCoreBundleVersion) = @"1938";
+static NSString * const CBLCoreBundleVersion NS_SWIFT_NAME(CascableCoreBundleVersion) = @"2054";